HW1/requirements.txt
numpy
biopython
seaborn

